requirements.txt
attrs==16.3.0
scipy==0.19.0
numpy==1.12.1
pandas==0.19.2
root-numpy==4.7.2
rootpy==0.9.0
